+++ versions/2.2.1/changelog.md
@@ -1,7 +1,7 @@
 Full Changelog: [*2.2.0* ➜ *2.2.1*](https://github.com/Furglitch/packwiz-kenkukraft/compare/2.2.0...2.2.1)
 
-
 <ins>**Swapped:**</ins>
 - *Multiplayer Server Pause* ➜ *Ready Player Fun*: MSP does not pause when running Chunky on an empty server.
   - Note: Chunky is not included in this pack by default, but can be added via your modpack/server manager. It's recommended to disable Distant Horizons while pre-generating chunks with Chunky.,
   - Added /setday command to change day counter as a fallback, in case player-offline pausing does not work as intended.
+- *Counter (Day/Death)* ➜ *PainterJS*: Reimplemented counter overlay with KubeJS script for better compatibility with /setday command.